+++ node/tests/SharedTests.ts
@@ -4384,7 +4384,7 @@ export function runBaseTests(config: {
                 expect(await client.zdiffstore(key4, [key1, key2])).toEqual(2);
                 const result1 = await client.zrangeWithScores(key4, {
                     start: 0,
-                    stop: -1,
+                    end: -1,
                 });
                 const expected1 = { one: 1.0, three: 3.0 };
                 expect(compareMaps(result1, expected1)).toBe(true);
@@ -4398,7 +4398,7 @@ export function runBaseTests(config: {
                 ).toEqual(1);
                 const result2 = await client.zrangeWithScores(key4, {
                     start: 0,
-                    stop: -1,
+                    end: -1,
                 });
                 expect(compareMaps(result2, { four: 4.0 })).toBe(true);
 
@@ -4407,7 +4407,7 @@ export function runBaseTests(config: {
                 ).toEqual(0);
                 const result3 = await client.zrangeWithScores(key4, {
                     start: 0,
-                    stop: -1,
+                    end: -1,
                 });
                 expect(compareMaps(result3, {})).toBe(true);
 
@@ -4416,7 +4416,7 @@ export function runBaseTests(config: {
                 ).toEqual(0);
                 const result4 = await client.zrangeWithScores(key4, {
                     start: 0,
-                    stop: -1,
+                    end: -1,
                 });
                 expect(compareMaps(result4, {})).toBe(true);
 
@@ -4468,7 +4468,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -4496,7 +4496,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -4524,7 +4524,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -4550,7 +4550,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -4575,7 +4575,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
         const membersScores1 = { one: 1.0, two: 2.0 };
         const membersScores2 = { one: 1.5, two: 2.5, three: 3.5 };
@@ -4611,7 +4611,7 @@ export function runBaseTests(config: {
         const key2 = "{testKey}:2-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
         const membersScores1 = { one: 1.0, two: 2.0 };
 
@@ -4780,12 +4780,13 @@ export function runBaseTests(config: {
                 const membersScores = { one: 1, two: 2, three: 3 };
                 expect(await client.zadd(key, membersScores)).toEqual(3);
 
-                expect(await client.zrange(key, { start: 0, stop: 1 })).toEqual(
-                    ["one", "two"],
-                );
+                expect(await client.zrange(key, { start: 0, end: 1 })).toEqual([
+                    "one",
+                    "two",
+                ]);
                 const result = await client.zrangeWithScores(key, {
                     start: 0,
-                    stop: -1,
+                    end: -1,
                 });
 
                 expect(
@@ -4798,15 +4799,15 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(
                         Buffer.from(key),
-                        { start: 0, stop: 1 },
+                        { start: 0, end: 1 },
                         { reverse: true, decoder: Decoder.Bytes },
                     ),
                 ).toEqual([Buffer.from("three"), Buffer.from("two")]);
-                expect(await client.zrange(key, { start: 3, stop: 1 })).toEqual(
+                expect(await client.zrange(key, { start: 3, end: 1 })).toEqual(
                     [],
                 );
                 expect(
-                    await client.zrangeWithScores(key, { start: 3, stop: 1 }),
+                    await client.zrangeWithScores(key, { start: 3, end: 1 }),
                 ).toEqual({});
             }, protocol);
         },
@@ -4824,13 +4825,13 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(key, {
                         start: InfBoundary.NegativeInfinity,
-                        stop: { value: 3, isInclusive: false },
+                        end: { value: 3, isInclusive: false },
                         type: "byScore",
                     }),
                 ).toEqual(["one", "two"]);
                 const result = await client.zrangeWithScores(Buffer.from(key), {
                     start: InfBoundary.NegativeInfinity,
-                    stop: InfBoundary.PositiveInfinity,
+                    end: InfBoundary.PositiveInfinity,
                     type: "byScore",
                 });
 
@@ -4846,7 +4847,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: { value: 3, isInclusive: false },
-                            stop: InfBoundary.NegativeInfinity,
+                            end: InfBoundary.NegativeInfinity,
                             type: "byScore",
                         },
                         { reverse: true },
@@ -4858,7 +4859,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: InfBoundary.PositiveInfinity,
+                            end: InfBoundary.PositiveInfinity,
                             limit: { offset: 1, count: 2 },
                             type: "byScore",
                         },
@@ -4871,7 +4872,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: { value: 3, isInclusive: false },
+                            end: { value: 3, isInclusive: false },
                             type: "byScore",
                         },
                         { reverse: true },
@@ -4881,7 +4882,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(key, {
                         start: InfBoundary.PositiveInfinity,
-                        stop: { value: 3, isInclusive: false },
+                        end: { value: 3, isInclusive: false },
                         type: "byScore",
                     }),
                 ).toEqual([]);
@@ -4891,7 +4892,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: { value: 3, isInclusive: false },
+                            end: { value: 3, isInclusive: false },
                             type: "byScore",
                         },
                         { reverse: true },
@@ -4901,7 +4902,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeWithScores(key, {
                         start: InfBoundary.PositiveInfinity,
-                        stop: { value: 3, isInclusive: false },
+                        end: { value: 3, isInclusive: false },
                         type: "byScore",
                     }),
                 ).toEqual({});
@@ -4921,7 +4922,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(Buffer.from(key), {
                         start: InfBoundary.NegativeInfinity,
-                        stop: { value: Buffer.from("c"), isInclusive: false },
+                        end: { value: Buffer.from("c"), isInclusive: false },
                         type: "byLex",
                     }),
                 ).toEqual(["a", "b"]);
@@ -4931,7 +4932,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.PositiveInfinity,
-                            stop: InfBoundary.NegativeInfinity,
+                            end: InfBoundary.NegativeInfinity,
                             limit: { offset: 1, count: 2 },
                             type: "byLex",
                         },
@@ -4944,7 +4945,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: { value: "c", isInclusive: false },
-                            stop: InfBoundary.NegativeInfinity,
+                            end: InfBoundary.NegativeInfinity,
                             type: "byLex",
                         },
                         { reverse: true },
@@ -4956,7 +4957,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: { value: "c", isInclusive: false },
+                            end: { value: "c", isInclusive: false },
                             type: "byLex",
                         },
                         { reverse: true },
@@ -4966,7 +4967,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(key, {
                         start: InfBoundary.PositiveInfinity,
-                        stop: { value: "c", isInclusive: false },
+                        end: { value: "c", isInclusive: false },
                         type: "byLex",
                     }),
                 ).toEqual([]);
@@ -4989,13 +4990,13 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, Buffer.from(key), {
                         start: 0,
-                        stop: 1,
+                        end: 1,
                     }),
                 ).toEqual(2);
                 expect(
                     await client.zrange(destkey, {
                         start: 0,
-                        stop: -1,
+                        end: -1,
                     }),
                 ).toEqual(["one", "two"]);
 
@@ -5003,7 +5004,7 @@ export function runBaseTests(config: {
                     await client.zrangeStore(
                         Buffer.from(destkey),
                         key,
-                        { start: 0, stop: 1 },
+                        { start: 0, end: 1 },
                         true,
                     ),
                 ).toEqual(2);
@@ -5012,7 +5013,7 @@ export function runBaseTests(config: {
                         destkey,
                         {
                             start: 0,
-                            stop: -1,
+                            end: -1,
                         },
                         { reverse: true },
                     ),
@@ -5021,7 +5022,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: 3,
-                        stop: 1,
+                        end: 1,
                     }),
                 ).toEqual(0);
             }, protocol);
@@ -5042,14 +5043,14 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.NegativeInfinity,
-                        stop: { value: 3, isInclusive: false },
+                        end: { value: 3, isInclusive: false },
                         type: "byScore",
                     }),
                 ).toEqual(2);
                 expect(
                     await client.zrange(destkey, {
                         start: 0,
-                        stop: -1,
+                        end: -1,
                     }),
                 ).toEqual(["one", "two"]);
 
@@ -5059,7 +5060,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: { value: 3, isInclusive: false },
-                            stop: InfBoundary.NegativeInfinity,
+                            end: InfBoundary.NegativeInfinity,
                             type: "byScore",
                         },
                         true,
@@ -5070,7 +5071,7 @@ export function runBaseTests(config: {
                         destkey,
                         {
                             start: 0,
-                            stop: -1,
+                            end: -1,
                         },
                         { reverse: true },
                     ),
@@ -5079,7 +5080,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.NegativeInfinity,
-                        stop: InfBoundary.PositiveInfinity,
+                        end: InfBoundary.PositiveInfinity,
                         limit: { offset: 1, count: 2 },
                         type: "byScore",
                     }),
@@ -5087,7 +5088,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(destkey, {
                         start: 0,
-                        stop: -1,
+                        end: -1,
                     }),
                 ).toEqual(["two", "three"]);
 
@@ -5097,7 +5098,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: { value: 3, isInclusive: false },
+                            end: { value: 3, isInclusive: false },
                             type: "byScore",
                         },
                         true,
@@ -5107,7 +5108,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.PositiveInfinity,
-                        stop: { value: 3, isInclusive: false },
+                        end: { value: 3, isInclusive: false },
                         type: "byScore",
                     }),
                 ).toEqual(0);
@@ -5129,21 +5130,21 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.NegativeInfinity,
-                        stop: { value: Buffer.from("c"), isInclusive: false },
+                        end: { value: Buffer.from("c"), isInclusive: false },
                         type: "byLex",
                     }),
                 ).toEqual(2);
                 expect(
                     await client.zrange(destkey, {
                         start: 0,
-                        stop: -1,
+                        end: -1,
                     }),
                 ).toEqual(["a", "b"]);
 
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.NegativeInfinity,
-                        stop: InfBoundary.PositiveInfinity,
+                        end: InfBoundary.PositiveInfinity,
                         limit: { offset: 1, count: 2 },
                         type: "byLex",
                     }),
@@ -5151,7 +5152,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(destkey, {
                         start: 0,
-                        stop: -1,
+                        end: -1,
                     }),
                 ).toEqual(["b", "c"]);
 
@@ -5161,7 +5162,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: { value: "c", isInclusive: false },
-                            stop: InfBoundary.NegativeInfinity,
+                            end: InfBoundary.NegativeInfinity,
                             type: "byLex",
                         },
                         true,
@@ -5172,7 +5173,7 @@ export function runBaseTests(config: {
                         destkey,
                         {
                             start: 0,
-                            stop: -1,
+                            end: -1,
                         },
                         { reverse: true },
                     ),
@@ -5184,7 +5185,7 @@ export function runBaseTests(config: {
                         key,
                         {
                             start: InfBoundary.NegativeInfinity,
-                            stop: { value: "c", isInclusive: false },
+                            end: { value: "c", isInclusive: false },
                             type: "byLex",
                         },
                         true,
@@ -5194,7 +5195,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, key, {
                         start: InfBoundary.PositiveInfinity,
-                        stop: { value: "c", isInclusive: false },
+                        end: { value: "c", isInclusive: false },
                         type: "byLex",
                     }),
                 ).toEqual(0);
@@ -5215,14 +5216,14 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(nonExistingKey, {
                         start: 0,
-                        stop: 1,
+                        end: 1,
                     }),
                 ).toEqual([]);
 
                 expect(
                     await client.zrangeWithScores(nonExistingKey, {
                         start: 0,
-                        stop: 1,
+                        end: 1,
                     }),
                 ).toEqual({});
 
@@ -5230,11 +5231,11 @@ export function runBaseTests(config: {
                 expect(await client.set(key, "value")).toEqual("OK");
 
                 await expect(
-                    client.zrange(key, { start: 0, stop: 1 }),
+                    client.zrange(key, { start: 0, end: 1 }),
                 ).rejects.toThrow();
 
                 await expect(
-                    client.zrangeWithScores(key, { start: 0, stop: 1 }),
+                    client.zrangeWithScores(key, { start: 0, end: 1 }),
                 ).rejects.toThrow();
 
                 // test zrangeStore - added in version 6.2.0
@@ -5244,13 +5245,13 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeStore(destkey, nonExistingKey, {
                         start: 0,
-                        stop: 1,
+                        end: 1,
                     }),
                 ).toEqual(0);
 
                 // test against a non-sorted set - throw RequestError
                 await expect(
-                    client.zrangeStore(destkey, key, { start: 0, stop: 1 }),
+                    client.zrangeStore(destkey, key, { start: 0, end: 1 }),
                 ).rejects.toThrow();
             }, protocol);
         },
@@ -5264,7 +5265,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -5311,7 +5312,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
 
         const membersScores1 = { one: 1.0, two: 2.0 };
@@ -5335,7 +5336,7 @@ export function runBaseTests(config: {
         const key3 = "{testKey}:3-" + uuidv4();
         const range = {
             start: 0,
-            stop: -1,
+            end: -1,
         };
         const membersScores1 = { one: 1.0, two: 2.0 };
         const membersScores2 = { one: 2.0, two: 3.0, three: 4.0 };
@@ -8803,7 +8804,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(4);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(searchResult);
 
                 // order search result
@@ -8825,7 +8826,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(4);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(searchResult);
 
                 // order and query all extra data
@@ -8871,7 +8872,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(1);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual([members[0]]);
 
                 // test search by box, unit: meters, from member, with distance
@@ -8905,7 +8906,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrangeWithScores(
                         key2,
-                        { start: 0, stop: -1 },
+                        { start: 0, end: -1 },
                         { reverse: true },
                     ),
                 ).toEqual({
@@ -8946,7 +8947,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(2);
                 expect(
-                    await client.zrangeWithScores(key2, { start: 0, stop: -1 }),
+                    await client.zrangeWithScores(key2, { start: 0, end: -1 }),
                 ).toEqual({
                     Palermo: 3479099956230698,
                     edge1: 3479273021651468,
@@ -8972,7 +8973,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(1);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(searchResult);
 
                 // test search by radius, units: feet, from member
@@ -8995,7 +8996,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(2);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(searchResult);
 
                 // Test search by radius, unit: meters, from member
@@ -9020,7 +9021,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(
                         key2,
-                        { start: 0, stop: -1 },
+                        { start: 0, end: -1 },
                         { reverse: true },
                     ),
                 ).toEqual(searchResult);
@@ -9065,7 +9066,7 @@ export function runBaseTests(config: {
                 expect(
                     await client.zrange(
                         key2,
-                        { start: 0, stop: -1 },
+                        { start: 0, end: -1 },
                         { reverse: true },
                     ),
                 ).toEqual(searchResult);
@@ -9099,7 +9100,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(2);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(members.slice(0, 2));
 
                 // Test search by radius, unit: kilometers, from a geospatial data, with limited ANY count to 1
@@ -9132,7 +9133,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(1);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual([searchResult[0][0]]);
 
                 // no members within the area
@@ -9781,7 +9782,7 @@ export function runBaseTests(config: {
                     ),
                 ).toEqual(4);
                 expect(
-                    await client.zrange(key2, { start: 0, stop: -1 }),
+                    await client.zrange(key2, { start: 0, end: -1 }),
                 ).toEqual(searchResult);
             }, protocol);
         },