+++ node/tests/TestUtilities.ts
@@ -1013,14 +1013,14 @@ export async function transactionTest(
 
     baseTransaction.zscore(key8, "member2");
     responseData.push(['zscore(key8, "member2")', 3.0]);
-    baseTransaction.zrange(key8, { start: 0, stop: -1 });
+    baseTransaction.zrange(key8, { start: 0, end: -1 });
     responseData.push([
-        "zrange(key8, { start: 0, stop: -1 })",
+        "zrange(key8, { start: 0, end: -1 })",
         ["member2", "member3", "member4", "member5"],
     ]);
-    baseTransaction.zrangeWithScores(key8, { start: 0, stop: -1 });
+    baseTransaction.zrangeWithScores(key8, { start: 0, end: -1 });
     responseData.push([
-        "zrangeWithScores(key8, { start: 0, stop: -1 })",
+        "zrangeWithScores(key8, { start: 0, end: -1 })",
         { member2: 3, member3: 3.5, member4: 4, member5: 5 },
     ]);
     baseTransaction.zadd(key12, { one: 1, two: 2 });
@@ -1055,9 +1055,9 @@ export async function transactionTest(
     responseData.push(["zadd(key13, { one: 1, two: 2, three: 3.5 })", 3]);
 
     if (gte(version, "6.2.0")) {
-        baseTransaction.zrangeStore(key8, key8, { start: 0, stop: -1 });
+        baseTransaction.zrangeStore(key8, key8, { start: 0, end: -1 });
         responseData.push([
-            "zrangeStore(key8, key8, { start: 0, stop: -1 })",
+            "zrangeStore(key8, key8, { start: 0, end: -1 })",
             4,
         ]);
         baseTransaction.zdiff([key13, key12]);